+++ apps/dashboard/src/@/utils/faucet.ts
@@ -7,6 +7,8 @@ const customClaimAmounts: Record<number, number> = {
   531050104: 5,
   // Arc testnet
   5042002: 1,
+  // Injective testnet
+  1439: 1,
 };
 
 const defaultClaimAmount = 0.01;