+++ packages/thirdweb/src/react/web/ui/ConnectWallet/Modal/AllWalletsUI.tsx
@@ -8,6 +8,7 @@ import walletInfos from "../../../../../wallets/__generated__/wallet-infos.js";
 import { NON_SEARCHABLE_WALLETS } from "../../../../../wallets/constants.js";
 import { createWallet } from "../../../../../wallets/create-wallet.js";
 import type { Wallet } from "../../../../../wallets/interfaces/wallet.js";
+import type { WalletId } from "../../../../../wallets/wallet-types.js";
 import { useCustomTheme } from "../../../../core/design-system/CustomThemeProvider.js";
 import { iconSize, spacing } from "../../../../core/design-system/index.js";
 import { useSetSelectionData } from "../../../providers/wallet-ui-states-provider.js";
@@ -34,15 +35,16 @@ function AllWalletsUI(props: {
   recommendedWallets: Wallet[] | undefined;
   connectLocale: ConnectLocale;
   disableSelectionDataReset?: boolean;
+  walletIdsToHide?: WalletId[];
 }) {
   const { itemsToShow, lastItemRef } = useShowMore<HTMLLIElement>(10, 10);
   const setSelectionData = useSetSelectionData();
 
   const walletList = useMemo(() => {
-    return walletInfos.filter(
-      (info) => !NON_SEARCHABLE_WALLETS.includes(info.id),
-    );
-  }, []);
+    return walletInfos
+      .filter((info) => !NON_SEARCHABLE_WALLETS.includes(info.id))
+      .filter((info) => !props.walletIdsToHide?.includes(info.id));
+  }, [props.walletIdsToHide]);
 
   const fuseInstance = useMemo(() => {
     return new Fuse(walletList, {